Switch to use-package's `:ensure` for pkg download
A recently merged feature to use-package allows checking the existence
of a package and invoking installation if it's missing.  Switching to
this method eliminates the need for keeping a list of packages and
also removes the necessity of bootstrapping while possibly taking a hit
to launch time by checking the existence of packages.