Emacs 24.3: goodbye ns-toggle-fullscreen and a bug
Full screen mode on Mac OS X is now implemented in the native way
introduced in Lion.  It's not necessarily better than
'ns-toggle-fullscreen but that patch is gone.  It's time to get used to
more OS X native way of full screen support.

24.3 seems to fix the drawing but that would add an empty column to
where right fringe was.  Remove our workaround for that.

Update the README, changing the Emacs version to 24.3 and OS X to
10.8.3.