Use org-mode from its elpa repository
To get the latest and greatest of Org, use it via `use-package' and source it
from orgmode.org ELPA repository.

`electric-indent-mode' is a global minor mode and there's no elegant way to
disable it when in org-mode.  As a remedy, map RET to `newline-and-indent' in
modes invoking `prog-mode-hook' which is almost all the modes I need auto
indentation.